Make Single Camera Session for Both Preview/Capture
Issue Detailed: Capture is creating a new camera session leading to
restart of camera sensor in HAL, Due to the restart first few frames of
camera are over exposed and when JPEG Encoded it is leading to yellow
frames.

Issue Fixed: Made Single Camera Session to be used for both
preview/capture which will avoid restart of camera sensor and the jpeg
encoding is proper.

Tracked-On: OAM-126049
Signed-off-by: NaveenVenturi1203 <venturi.naveen@intel.com>